block file upload inside shareview
